Move some init code around
