+++ client/components/App.vue
@@ -1,6 +1,6 @@
 <template>
 	<div id="viewport" :class="viewportClasses" role="tablist">
-		<Sidebar :overlay="$refs.overlay" />
+		<Sidebar v-if="$store.state.appLoaded" :overlay="$refs.overlay" />
 		<div id="sidebar-overlay" ref="overlay" @click="$store.commit('sidebarOpen', false)" />
 		<article id="windows">
 			<router-view ref="window"></router-view>
@@ -11,6 +11,8 @@
 
 <script>
 const throttle = require("lodash/throttle");
+const constants = require("../js/constants");
+const storage = require("../js/localStorage");
 
 import Sidebar from "./Sidebar.vue";
 import ImageViewer from "./ImageViewer.vue";
@@ -25,11 +27,14 @@ export default {
 		viewportClasses() {
 			return {
 				notified: this.$store.getters.highlightCount > 0,
-				"menu-open": this.$store.state.sidebarOpen,
+				"menu-open": this.$store.state.appLoaded && this.$store.state.sidebarOpen,
 				"menu-dragging": this.$store.state.sidebarDragging,
 				"userlist-open": this.$store.state.userlistOpen,
 			};
 		},
+	},
+	created() {
+		this.prepareOpenStates();
 	},
 	mounted() {
 		// Make a single throttled resize listener available to all components
@@ -59,6 +64,25 @@ export default {
 			const tommorow = new Date(today.getFullYear(), today.getMonth(), today.getDate() + 1);
 
 			return tommorow - today;
+		},
+		prepareOpenStates() {
+			const viewportWidth = window.outerWidth;
+			let isUserlistOpen = storage.get("thelounge.state.userlist");
+
+			if (viewportWidth > constants.mobileViewportPixels) {
+				this.$store.commit(
+					"sidebarOpen",
+					storage.get("thelounge.state.sidebar") !== "false"
+				);
+			}
+
+			// If The Lounge is opened on a small screen (less than 1024px), and we don't have stored
+			// user list state, close it by default
+			if (viewportWidth >= 1024 && isUserlistOpen !== "true" && isUserlistOpen !== "false") {
+				isUserlistOpen = "true";
+			}
+
+			this.$store.commit("userlistOpen", isUserlistOpen === "true");
 		},
 	},
 };
